Stash: Refactor Node ID field retrieval into compiler context.
